proper linking to SFML
