+++ .idea/libraries/workspace.xml
@@ -1,0 +1,8 @@
+<component name="libraryTable">
+  <library name="workspace">
+    <CLASSES />
+    <SOURCES>
+      <root url="file://$USER_HOME$" />
+    </SOURCES>
+  </library>
+</component>
