--- simulator/basicrobot.cpp
@@ -1,17 +1,0 @@
-#include "basicrobot.h"
-#include "drawable_robot.h"
-
-BasicRobot::BasicRobot(double x, double y, double angle) {
-	x_pos = x;
-	y_pos = y;
-	angle = angle;
-
-	width = 30;
-	length = 30;
-
-	wheel_radius = 0.05;
-	left_wheel_velocity = 0;
-	right_wheel_velocity = 0;
-	max_wheel_velocity = 8;
-	wheel_acceleration = 1;
-}
