+++ simulator/include/simulator/basicrobot.h
@@ -1,0 +1,11 @@
+#ifndef basicrobot_h
+#define basicrobot_h
+
+#include <simulator/robot.h>
+#include <simulator/drawable_robot.h>
+
+class BasicRobot: public Robot { //basic robot for testing purposes
+	public: BasicRobot(double x, double y, double angle);
+};
+
+#endif
