+++ simulator/include/simulator/drawable_robot.h
