+++ simulator/include/simulator/hardware_interaction.h
