+++ simulator/include/simulator/sensors/abstract_sensor.h
