+++ simulator/src/CMakeLists.txt
@@ -1,0 +1,11 @@
+
+
+
+
+
+
+
+
+
+add_executable(sim simulator/sim.cpp)
+target_link_libraries(sim ${SFML_LIBRARIES} ${SFML_DEPENDENCIES})
