+++ simulator/src/simulator/sensors/distance_sensor.cpp
@@ -1,4 +1,4 @@
-#include "abstract_sensor.h"
+#include "../../../include/simulator/sensors/abstract_sensor.h"
 
 class DistanceSensor : public AbstractSensor {
     public:
