+++ strategy/include/strategy/slammywhammy.h
@@ -1,0 +1,8 @@
+//
+// Created by jason on 9/17/20.
+//
+
+#ifndef ROBOWRESTLING_SOFTWARE_SLAMMYWHAMMY_H
+#define ROBOWRESTLING_SOFTWARE_SLAMMYWHAMMY_H
+
+#endif //ROBOWRESTLING_SOFTWARE_SLAMMYWHAMMY_H
