+++ strategy/include/strategy/strategy.h
