+++ strategy/src/slammywhammy.cpp
@@ -1,0 +1,6 @@
+#include <strategy/slammywhammy.h>
+#include <iostream>
+
+int main() {
+  std::cout << "hello world" << std::endl;
+}
